Refactored HeuristicAsyncTask
